Refactor table stuff into directories
