+++ acs-admin/src/pages/ConfigDB/Applications/TableData/applicationListColumns.ts
