+++ acs-admin/src/pages/ConfigDB/Objects/TableData/objectMemberOfListColumns.ts
@@ -5,15 +5,16 @@
 import type {ColumnDef} from '@tanstack/vue-table'
 import {h} from 'vue'
 import DataTableColumnHeader from '@/components/ui/data-table/DataTableColumnHeader.vue'
-import SubclassOfDropdown from "@pages/ConfigDB/Objects/SubclassOfDropdown.vue";
+import MemberOfDropdown from "./MemberOfDropdown.vue";
 
-export interface SubclassOfMapping {
+export interface MemberOfMapping {
     uuid: string
     name: string
     class: {
         uuid: string
         name: string
     }
+    direct: string
     originalObject: {
         uuid: string
         name: string
@@ -24,7 +25,7 @@ export interface SubclassOfMapping {
     }
 }
 
-export const subclassOfColumns: ColumnDef<SubclassOfMapping>[] = [{
+export const memberOfColumns: ColumnDef<MemberOfMapping>[] = [{
     accessorKey: 'name',
     header: ({column}) => h(DataTableColumnHeader, {
         column,
@@ -48,15 +49,28 @@ export const subclassOfColumns: ColumnDef<SubclassOfMapping>[] = [{
     }),
     cell: ({row}) => {
         return h('div', {class: 'max-w-[500px] truncate'}, [
-            h('div', {class: 'max-w-[500px] truncate font-medium'}, row.getValue('class')),
+            h('div', {class: 'max-w-[500px] truncate'}, row.getValue('class')),
             h('div', {class: 'max-w-[500px] truncate text-gray-400'}, row.original.class?.uuid ?? "UNKNOWN")
         ])
     },
     filterFn: (row, id, value) => {
         return value.includes(row.getValue(id))
     },
+}, {
+    accessorKey: 'direct',
+    accessorFn: (item) => item.direct,
+    header: ({column}) => h(DataTableColumnHeader, {
+        column,
+        title: 'Direct'
+    }),
+    cell: ({row}) => {
+        return h('div', {class: 'max-w-[500px] truncate font-medium'}, row.getValue('direct'))
+    },
+    filterFn: (row, id, value) => {
+        return value.includes(row.getValue(id))
+    },
 },
-    {
-        id: 'actions',
-        cell: ({row}) => h(SubclassOfDropdown, {row}),
-    }]
+{
+    id: 'actions',
+    cell: ({row}) => h(MemberOfDropdown, {row}),
+}]
